Dev: add Scatters to Facet
