+++ charts/facetChart/FacetChart.stories.tsx
@@ -2,8 +2,9 @@ import * as React from "react"
 import "site/client/owid.scss"
 import "charts/core/chart.scss"
 import { FacetChart } from "./FacetChart"
-import { basicGdpChart } from "charts/test/samples"
+import { basicGdpChart, basicScatter } from "charts/test/samples"
 import { Bounds } from "charts/utils/Bounds"
+import { ChartTypeName } from "charts/core/ChartTypes"
 
 export default {
     title: "FacetChart",
@@ -14,14 +15,17 @@ export default {
 }
 
 export const Default = (args: any) => {
-    const chart = basicGdpChart()
+    const chartType: ChartTypeName = args.chartTypeName || "LineChart"
+    const chart = chartType.includes("Scatter")
+        ? basicScatter()
+        : basicGdpChart()
     const bounds = new Bounds(0, 0, 640, 480)
 
     return (
         <svg width={640} height={480}>
             <FacetChart
                 number={args.number || 4}
-                chartTypeName={args.chartTypeName || "LineChart"}
+                chartTypeName={chartType}
                 chart={chart}
                 bounds={bounds}
             />
