+++ charts/test/samples.ts
@@ -37,3 +37,15 @@ Germany,2003,120,1,256`) as any
     chartConfig.table.columnsBySlug.get("population")!.spec.owidVariableId = 100
     return chartConfig
 }
+
+export const basicScatter = () => {
+    const chartRuntime = basicGdpChart()
+    const script = chartRuntime.props
+    script.type = "ScatterPlot"
+    chartRuntime.yAxisOptions.min = 0
+    chartRuntime.yAxisOptions.max = 500
+    chartRuntime.xAxisOptions.min = 0
+    chartRuntime.xAxisOptions.max = 500
+    script.dimensions.push({ variableId: 100, property: "x", display: {} })
+    return chartRuntime
+}
